feat: use typed Market constants (P6-2 done)
- Removed int32() casts from market parameters
- Typed constants: constant.Market_US, constant.Market_HK
- Demo now uses compile-time type safety

+++ examples/06_kline_single/main.go
@@ -22,11 +22,11 @@ func main() {
 		log.Fatalf("Connect failed: %v", err)
 	}
 
-	if err := client.Subscribe(context.Background(), cli, int32(constant.Market_US), "NVDA", []constant.SubType{constant.SubType_K_Day}); err != nil {
+	if err := client.Subscribe(context.Background(), cli, constant.Market_US, "NVDA", []constant.SubType{constant.SubType_K_Day}); err != nil {
 		log.Fatalf("Subscribe failed: %v", err)
 	}
 
-	klines, err := client.GetKLines(context.Background(), cli, int32(constant.Market_US), "NVDA", int32(constant.KLType_K_Day), 10)
+	klines, err := client.GetKLines(context.Background(), cli, constant.Market_US, "NVDA", constant.KLType_K_Day, 10)
 	if err != nil {
 		log.Fatalf("GetKLines failed: %v", err)
 	}
