+++ examples/08_orderbook_req/main.go
@@ -22,11 +22,11 @@ func main() {
 		log.Fatalf("Connect failed: %v", err)
 	}
 
-	if err := client.Subscribe(context.Background(), cli, int32(constant.Market_US), "NVDA", []constant.SubType{constant.SubType_OrderBook}); err != nil {
+	if err := client.Subscribe(context.Background(), cli, constant.Market_US, "NVDA", []constant.SubType{constant.SubType_OrderBook}); err != nil {
 		log.Fatalf("Subscribe failed: %v", err)
 	}
 
-	book, err := client.GetOrderBook(context.Background(), cli, int32(constant.Market_US), "NVDA", 10)
+	book, err := client.GetOrderBook(context.Background(), cli, constant.Market_US, "NVDA", 10)
 	if err != nil {
 		log.Fatalf("GetOrderBook failed: %v", err)
 	}
